refactor: validation 삭제 및 각 repository에 통합
