+++ src/main/java/com/sparta/todo/domain/comment/repository/CommentRepository.java
@@ -2,6 +2,7 @@ package com.sparta.todo.domain.comment.repository;
 
 import com.sparta.todo.domain.comment.entity.CommentEntity;
 import com.sparta.todo.domain.comment.model.Comment;
+import java.util.NoSuchElementException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 
@@ -10,6 +11,12 @@ import org.springframework.stereotype.Repository;
 public class CommentRepository {
 
     private final CommentJpaRepository commentJpaRepository;
+
+    public Comment findCommentBy(Long id) {
+        return Comment.from(commentJpaRepository.findById(id).orElseThrow(
+            () -> new NoSuchElementException("해당 댓글은 존재하지 않습니다.")
+        ));
+    }
 
     public void save(CommentEntity commentEntity) {
         commentJpaRepository.save(commentEntity);
