+++ src/main/java/com/sparta/todo/domain/comment/service/CommentService.java
@@ -6,8 +6,8 @@ import com.sparta.todo.domain.comment.entity.CommentEntity;
 import com.sparta.todo.domain.comment.model.Comment;
 import com.sparta.todo.domain.comment.repository.CommentRepository;
 import com.sparta.todo.domain.todo.model.Todo;
+import com.sparta.todo.domain.todo.repository.TodoRepository;
 import com.sparta.todo.domain.user.model.User;
-import com.sparta.todo.global.validation.Validation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -17,13 +17,13 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional
 public class CommentService {
 
-    private final Validation validation;
+    private final TodoRepository todoRepository;
     private final CommentRepository commentRepository;
 
     public CommentResponseDto createComment(User user, Long todoId,
         CommentRequestDto requestDto) {
 
-        Todo todo = validation.findTodoBy(todoId);
+        Todo todo = todoRepository.findTodoBy(todoId);
 
         CommentEntity commentEntity = new CommentEntity(requestDto, todo, user);
         commentRepository.save(commentEntity);
@@ -34,8 +34,8 @@ public class CommentService {
     public CommentResponseDto updateComment(User user, Long todoId, Long commentId,
         CommentRequestDto requestDto) {
 
-        Comment comment = validation.findCommentBy(commentId);
-        Todo todo = validation.findTodoBy(todoId);
+        Comment comment = commentRepository.findCommentBy(commentId);
+        Todo todo = todoRepository.findTodoBy(todoId);
 
         comment.validateBy(todo);
         comment.validateBy(user);
@@ -47,9 +47,9 @@ public class CommentService {
     }
 
     public void deleteComment(Long todoId, Long commentId) {
-        Comment comment = validation.findCommentBy(commentId);
+        Comment comment = commentRepository.findCommentBy(commentId);
 
-        validation.findTodoBy(todoId);
+        todoRepository.findTodoBy(todoId);
 
         commentRepository.delete(comment);
     }
