+++ src/main/java/com/sparta/todo/domain/todo/repository/TodoRepository.java
@@ -2,6 +2,7 @@ package com.sparta.todo.domain.todo.repository;
 
 import com.sparta.todo.domain.todo.entity.TodoEntity;
 import com.sparta.todo.domain.todo.model.Todo;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 
@@ -10,6 +11,12 @@ import org.springframework.stereotype.Repository;
 public class TodoRepository {
 
     private final TodoJpaRepository todoJpaRepository;
+
+    public Todo findTodoBy(Long id) {
+        return Todo.from(todoJpaRepository.findById(id).orElseThrow(() ->
+            new EntityNotFoundException("선택한 일정은 존재하지 않습니다.")
+        ));
+    }
 
     public void save(TodoEntity todoEntity) {
         todoJpaRepository.save(todoEntity);
