+++ src/main/java/com/sparta/todo/domain/user/repository/UserRepository.java
@@ -3,6 +3,7 @@ package com.sparta.todo.domain.user.repository;
 import com.sparta.todo.domain.user.entity.UserEntity;
 import com.sparta.todo.domain.user.model.User;
 import java.util.List;
+import java.util.NoSuchElementException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.dao.DuplicateKeyException;
 import org.springframework.stereotype.Repository;
@@ -11,6 +12,12 @@ import org.springframework.stereotype.Repository;
 @RequiredArgsConstructor
 public class UserRepository {
     private final UserJpaRepository userJpaRepository;
+
+    public User userBy(String userName) {
+        return User.from(userJpaRepository.findByUserName(userName).orElseThrow(
+            () -> new NoSuchElementException("사용자를 찾을 수 없습니다.")
+        ));
+    }
     
     public void save(UserEntity userEntity) {
         userJpaRepository.save(userEntity);
